test: add scopeds tests
