+++ stores/flutter_triple/test/src/widgets/scoped_listener_test.dart
@@ -1,250 +1,124 @@
-// ignore_for_file: type_annotate_public_apis, always_declare_return_types
-import 'dart:developer';
-
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_triple/flutter_triple.dart';
+import 'package:flutter_triple/src/stores/notifier_store.dart';
+import 'package:flutter_triple/src/widgets/scoped_listener.dart';
 
 void main() {
-  test('instance', () {
-    expect(
-      ScopedListener(
-        store: CounterStore(),
-        onState: (context, state) => log(state.toString()),
+  group('ScopedListener', () {
+    late TripleTeste store;
+    late ScopedListener scopedListener;
+
+    setUp(() {
+      store = TripleTeste();
+      scopedListener = ScopedListener<TripleTeste, TripleTesteError, int>(
+        store: store,
+        onState: (context, state) {},
         child: Container(),
-      ),
-      isA<ScopedListener>(),
-    );
-  });
-  test('throw assert when no have onState, onLoading, onError', () {
-    expect(
-      ScopedListener(
-        store: CounterStore(),
-        onState: (context, state) => log(state.toString()),
-        child: Container(),
-      ),
-      isA<ScopedListener>(),
-    );
-    expect(
-      () => ScopedListener(
-        store: CounterStore(),
-        child: Container(),
-      ),
-      throwsAssertionError,
-    );
-  });
-  test('''throw assert when have distinct but don't have onState''', () {
-    expect(
-      ScopedListener(
-        store: CounterStore(),
-        onState: (context, state) => log(state.toString()),
-        child: Container(),
-      ),
-      isA<ScopedListener>(),
-    );
-    expect(
-      () => ScopedListener(
-        store: CounterStore(),
-        distinct: (s) => s,
-        onLoading: (context, isLoading) => log(isLoading.toString()),
-        child: Container(),
-      ),
-      throwsAssertionError,
-    );
-  });
-  test('''throw assert when have filter but don't have onState''', () {
-    expect(
-      ScopedListener(
-        store: CounterStore(),
-        onState: (context, state) => log(state.toString()),
-        child: Container(),
-      ),
-      isA<ScopedListener>(),
-    );
-    expect(
-      () => ScopedListener(
-        store: CounterStore(),
-        filter: (s) => true,
-        onLoading: (context, isLoading) => log(isLoading.toString()),
-        child: Container(),
-      ),
-      throwsAssertionError,
-    );
-  });
-  testWidgets('test change state', (WidgetTester tester) async {
-    final counter = CounterStore();
-    await tester.pumpWidget(
-      MaterialApp(
-        home: TestCounterPage(
-          store: counter,
-        ),
-      ),
-    );
-    final buttonFinder = find.byType(FloatingActionButton);
-    expect(buttonFinder, findsOneWidget);
-    expect(find.text('0'), findsOneWidget);
-    await tester.tap(buttonFinder);
-    await tester.pump();
-    expect(find.text('1'), findsOneWidget);
-    await tester.tap(buttonFinder);
-    await tester.pump();
-    expect(find.text('2'), findsOneWidget);
-  });
+      );
+    });
 
-  testWidgets('test change state store directly', (WidgetTester tester) async {
-    final counter = CounterStore();
-    await tester.pumpWidget(
-      MaterialApp(
-        home: TestCounterPage(
-          store: counter,
-        ),
-      ),
-    );
-    expect(find.text('0'), findsOneWidget);
-    counter.increment();
-    await tester.pump();
-    expect(find.text('1'), findsOneWidget);
-    counter.increment();
-    await tester.pump();
-    expect(find.text('2'), findsOneWidget);
-  });
-
-  testWidgets('test change state filter', (WidgetTester tester) async {
-    final counter = CounterStore();
-    await tester.pumpWidget(
-      MaterialApp(
-        home: TestCounterPage(
-          store: counter,
-          withFilter: true,
-        ),
-      ),
-    );
-    expect(find.text('0'), findsOneWidget);
-    counter.increment();
-    await tester.pump();
-    expect(find.text('1'), findsOneWidget);
-    counter.increment();
-    await tester.pump();
-    expect(find.text('1'), findsOneWidget);
-    counter.increment();
-    await tester.pump();
-    expect(find.text('3'), findsOneWidget);
-  });
-  testWidgets('test change state distinct', (WidgetTester tester) async {
-    final counter = CounterStore();
-    final rebuild = [];
-    await tester.pumpWidget(
-      MaterialApp(
-        home: TestCounterPage(
-          store: counter,
-          onState: rebuild,
-        ),
-      ),
-    );
-
-    expect(find.text('0'), findsOneWidget);
-    expect(rebuild.length, 1);
-
-    counter.increment();
-    await tester.pump();
-    expect(find.text('1'), findsOneWidget);
-    expect(rebuild.length, 2);
-
-    counter.increment();
-    await tester.pump();
-    expect(find.text('2'), findsOneWidget);
-    expect(rebuild.length, 3);
-    counter.increment();
-
-    await tester.pump();
-    expect(find.text('3'), findsOneWidget);
-    expect(rebuild.length, 4);
-    counter.increment();
-
-    await tester.pump();
-    expect(find.text('3'), findsOneWidget);
-    expect(rebuild.length, 4);
-  });
-}
-
-class CounterState {
-  final int value;
-
-  CounterState(this.value);
-
-  operator +(int newValue) => CounterState(value + newValue);
-}
-
-class CounterStore extends NotifierStore<Exception, CounterState>
-    with MementoMixin {
-  CounterStore() : super(CounterState(0));
-
-  void increment() {
-    if (state.value > 2) {
-      update(CounterState(3));
-    } else {
-      update(state + 1);
-    }
-  }
-}
-
-class TestCounterPage extends StatelessWidget {
-  final CounterStore store;
-  final bool withFilter;
-  final List? onLoading;
-  final List? onState;
-  final List? onError;
-
-  const TestCounterPage({
-    Key? key,
-    required this.store,
-    this.withFilter = false,
-    this.onLoading,
-    this.onState,
-    this.onError,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        actions: [
-          IconButton(
-            onPressed: store.undo,
-            icon: const Icon(Icons.arrow_back_ios),
-          ),
-          IconButton(
-            onPressed: store.redo,
-            icon: const Icon(Icons.arrow_forward_ios),
-          ),
-        ],
-      ),
-      body: Center(
-        child: ScopedListener<CounterStore, Exception, CounterState>(
+    testWidgets('should throw an error if no listeners are defined',
+        (tester) async {
+      expect(
+        () => ScopedListener<TripleTeste, TripleTesteError, int>(
           store: store,
-          distinct: (state) => state.value,
-          filter: withFilter ? (state) => state.value != 2 : null,
-          onLoading: (context, isLoading) => onLoading?.add(0),
-          onState: (context, state) => onState?.add(0),
-          onError: (context, error) => onError?.add(0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              const Text('You have pushed the button this many times:'),
-              Text(
-                '${store.state.value}',
-                style: Theme.of(context).textTheme.headlineMedium,
-              ),
-            ],
-          ),
+          child: Container(),
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: store.increment,
-        tooltip: 'Increment',
-        backgroundColor: Colors.grey,
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
+        throwsAssertionError,
+      );
+    });
+
+    testWidgets(
+        'should throw an error if distinct is defined but onState is not',
+        (tester) async {
+      expect(
+        () => ScopedListener<TripleTeste, TripleTesteError, int>(
+          store: store,
+          distinct: (state) => state,
+          child: Container(),
+        ),
+        throwsAssertionError,
+      );
+    });
+
+    testWidgets('should throw an error if filter is defined but onState is not',
+        (tester) async {
+      expect(
+        () => ScopedListener<TripleTeste, TripleTesteError, int>(
+          store: store,
+          filter: (state) => true,
+          child: Container(),
+        ),
+        throwsAssertionError,
+      );
+    });
+
+    testWidgets('should render child widget', (tester) async {
+      await tester.pumpWidget(scopedListener);
+      expect(find.byType(Container), findsOneWidget);
+    });
+
+    testWidgets('should trigger onState when state changes', (tester) async {
+      var count = 0;
+      scopedListener = ScopedListener<TripleTeste, TripleTesteError, int>(
+        store: store,
+        onState: (context, state) {
+          count++;
+        },
+        child: Container(),
+      );
+
+      await tester.pumpWidget(scopedListener);
+      store.update(1);
+      await tester.pump();
+
+      expect(count, equals(1));
+    });
+
+    testWidgets(
+        'should not trigger onState when state changes but filter returns false',
+        (tester) async {
+      var count = 0;
+      scopedListener = ScopedListener<TripleTeste, TripleTesteError, int>(
+        store: store,
+        onState: (context, state) {
+          count++;
+        },
+        filter: (state) => state > 1,
+        child: Container(),
+      );
+
+      await tester.pumpWidget(scopedListener);
+      store.update(1);
+      await tester.pump();
+
+      expect(count, equals(0));
+    });
+
+    testWidgets('should trigger onError when error is thrown', (tester) async {
+      var count = 0;
+      scopedListener = ScopedListener<TripleTeste, TripleTesteError, int>(
+        store: store,
+        onError: (context, error) {
+          count++;
+        },
+        child: Container(),
+      );
+
+      await tester.pumpWidget(scopedListener);
+      store.setError(TripleTesteError(''));
+      await tester.pump();
+
+      expect(count, equals(1));
+    });
+  });
+}
+
+class TripleTeste extends NotifierStore<TripleTesteError, int> {
+  TripleTeste() : super(0);
+}
+
+class TripleTesteError {
+  final String message;
+  TripleTesteError(this.message);
 }
